+++ lib/src/geom/modifier/dodge.dart
@@ -1,9 +1,10 @@
 import 'dart:ui';
 
+import 'package:graphic/src/algebra/varset.dart';
+import 'package:graphic/src/coord/coord.dart';
 import 'package:graphic/src/dataflow/tuple.dart';
 import 'package:graphic/src/scale/discrete.dart';
 import 'package:graphic/src/scale/scale.dart';
-import 'package:graphic/src/algebra/varset.dart';
 
 import 'modifier.dart';
 
@@ -17,7 +18,7 @@ class DodgeModifier extends Modifier {
     this.symmetric,
   });
 
-  /// The dodge ratio to the descrete band for each group.
+  /// The dodge ratio to the discrete band for each group.
   ///
   /// If null, a default reciprocal of group counts is set.
   double? ratio;
@@ -29,83 +30,34 @@ class DodgeModifier extends Modifier {
   bool? symmetric;
 
   @override
-  bool operator ==(Object other) =>
+  bool equalTo(Object other) =>
       other is DodgeModifier &&
-      super == other &&
       ratio == other.ratio &&
       symmetric == other.symmetric;
-}
-
-/// The dodge geometory modifier.
-class DodgeGeomModifier extends GeomModifier {
-  DodgeGeomModifier(
-    this.ratio,
-    this.symmetric,
-    this.band,
-  );
-
-  /// The dodge ratio to the descrete band for each group.
-  final double ratio;
-
-  /// Whether the dodge will go both side around the original x or only positive
-  /// side.
-  final bool symmetric;
-
-  /// The band ratio of each value.
-  final double band;
 
   @override
-  void modify(AesGroups value) {
-    final bias = ratio * band;
-    var accumulated = 0.0;
+  void modify(AesGroups groups, Map<String, ScaleConv<dynamic, num>> scales,
+      AlgForm form, CoordConv coord, Offset origin) {
+    final xField = form.first[0];
+    final band = (scales[xField] as DiscreteScaleConv).band;
 
-    for (var group in value) {
+    final ratio = this.ratio ?? 1 / (groups.length);
+    final symmetric = this.symmetric ?? true;
+
+    final bias = ratio * band;
+    // If symmetric, negtively shifts half of the total bias.
+    var accumulated = symmetric ? -bias * (groups.length - 1) / 2 : 0.0;
+
+    for (var group in groups) {
       for (var aes in group) {
         final oldPosition = aes.position;
         aes.position = oldPosition
             .map(
-              (point) => Offset(point.dx + accumulated + bias, point.dy),
+              (point) => Offset(point.dx + accumulated, point.dy),
             )
             .toList();
       }
       accumulated += bias;
     }
-
-    if (symmetric) {
-      final symmetricBias = -accumulated / 2;
-      for (var group in value) {
-        for (var aes in group) {
-          final oldPosition = aes.position;
-          aes.position = oldPosition
-              .map(
-                (point) => Offset(point.dx + symmetricBias, point.dy),
-              )
-              .toList();
-        }
-      }
-    }
-  }
-}
-
-/// The dodge geometory modifier operator.
-class DodgeGeomModifierOp extends GeomModifierOp<DodgeGeomModifier> {
-  DodgeGeomModifierOp(Map<String, dynamic> params) : super(params);
-
-  @override
-  DodgeGeomModifier evaluate() {
-    final ratio = params['ratio'] as double?;
-    final symmetric = params['symmetric'] as bool;
-    final form = params['form'] as AlgForm;
-    final scales = params['scales'] as Map<String, ScaleConv>;
-    final groups = params['groups'] as AesGroups;
-
-    final xField = form.first[0];
-    final band = (scales[xField] as DiscreteScaleConv).band;
-
-    return DodgeGeomModifier(
-      ratio ?? 1 / (groups.length),
-      symmetric,
-      band,
-    );
   }
 }